+++ src/datasources/server_state_dao.go
@@ -22,3 +22,43 @@ type ServerStateDao interface {
 	// received this server's vote in the current term
 	VotedFor() (int64, int64)
 }
+
+// MakeInMemoryServerStateDao creates an instance of inMemoryServerStateDao for
+// use in unit tests outside of this package
+func MakeInMemoryServerStateDao(currentTerm,
+	votedFor int64) *inMemoryServerStateDao {
+	return &inMemoryServerStateDao{currentTerm: currentTerm, votedFor: votedFor}
+}
+
+// inMemoryServerStateDao implements an in-memory version of the ServerStateDao
+// interface. This implementation is useful for testing code
+type inMemoryServerStateDao struct {
+	currentTerm int64
+	votedFor    int64
+}
+
+func (s *inMemoryServerStateDao) CurrentTerm() int64 {
+	return s.currentTerm
+}
+
+func (s *inMemoryServerStateDao) UpdateTerm(serverTerm int64) error {
+	s.currentTerm = serverTerm
+	s.votedFor = -1
+	return nil
+}
+
+func (s *inMemoryServerStateDao) UpdateTermVotedFor(serverTerm int64,
+	serverID int64) error {
+	s.currentTerm = serverTerm
+	s.votedFor = serverID
+	return nil
+}
+
+func (s *inMemoryServerStateDao) UpdateVotedFor(serverID int64) error {
+	s.votedFor = serverID
+	return nil
+}
+
+func (s *inMemoryServerStateDao) VotedFor() (int64, int64) {
+	return s.currentTerm, s.votedFor
+}
