+++ src/domain/errors.go
@@ -12,6 +12,7 @@ const (
 	notImplementedErrFmt  = "method %s not implemented"
 	termInconsistencyErr  = "remote server term less than local term"
 	multipleLeadersErrFmt = "multiple leaders detected during term %d"
+	invalidIndexErrFmt    = "log entry index %d is invalid"
 
 	// Unit test errors
 	methodExpectedToFailErrFmt    = "method %s expected to fail"
